refactored chartStyler's entry component
