+++ src/components/chartStyler/index.tsx
@@ -1,10 +1,6 @@
-import { useAppDispatch, useDebounce } from "@/lib/hooks";
 import { TParetoChartData } from "../pareto/chart";
-import { rCombined } from "@/lib/reducers";
-import { chart } from "@/lib/utils";
 
-import LineTypeOption from "./typeSelector";
-import { useState } from "react";
+import Entry from "./entry";
 
 export default function ChartStyler(pp: { users: TParetoChartData["users"] }) {
   return (
@@ -15,46 +11,8 @@ export default function ChartStyler(pp: { users: TParetoChartData["users"] }) {
       }
     >
       {pp.users.map((user) => (
-        <Child key={user.id} {...user} />
+        <Entry key={user.id} {...user} />
       ))}
     </div>
   );
 }
-
-const Child = ({ id, name, chartStyle }: TParetoChartData["users"]["0"]) => {
-  const dispatch = useAppDispatch();
-  const [color, setColor] = useState(chart(chartStyle).color);
-
-  useDebounce(
-    () =>
-      dispatch(
-        rCombined.setChartStyle(
-          `${chartStyle?.at(0) ?? "0"}${color.slice(1)}`,
-          id
-        )
-      ),
-    [color]
-  );
-
-  return (
-    <div
-      key={id}
-      className="flex flex-col gap-2 items-center"
-      style={{ color }}
-    >
-      {name}
-
-      <div className="flex p-2 gap-2 justify-between">
-        <input
-          type="color"
-          className="border-0! cursor-pointer"
-          value={color}
-          onChange={(ev) => setColor(ev.target.value)}
-        />
-
-        <LineTypeOption {...{ chartStyle, id, setTo: "solid" }} />
-        <LineTypeOption {...{ chartStyle, id, setTo: "dashed" }} />
-      </div>
-    </div>
-  );
-};
